[MULTI-28] Testing: modify postgres collat pool testing to match schema

+++ indexer/packages/postgres/src/db/migrations/migration_files/20241220173012_create_collateral_pools_table.ts
@@ -3,7 +3,7 @@ import Knex from 'knex';
 export async function up(knex: Knex): Promise<void> {
   return knex.schema.createTable('collateral_pools', (table) => {
     table.integer('id').primary();
-    table.integer('maxCumulativeInsuranceFundDeltaPerBlock').notNullable();
+    table.bigInteger('maxCumulativeInsuranceFundDeltaPerBlock').notNullable();
     table.specificType('multiCollateralAssets', 'INTEGER[]').notNullable();
     table.integer('quoteAssetId').notNullable();
   });
